Task to calculate length of a 2D or 3D vector

+++ csharp-linear_algebra/3-magnitude/3-magnitude.cs
@@ -5,7 +5,7 @@ class VectorMath
 {
     public static double Magnitude(double[] vector)
     {
-        int dimensions = vector.Length;
+        Double dimensions = vector.Length;
 
         if (dimensions != 2 || dimensions != 3)
         {
